Remove contact title from API, fixtures and constants
Resolves #361

This partially removes title from contacts as it was removed from the front end some time ago. However, the field has been kept on the model as there is some it contains some migrated data and there's a small chance it might come back.

As if it comes back it would at least be with a reduced set of options, I think we're better off removing it from the constants, fixtures and API rather than keeping it just in case.